% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R, R/generics.R
\name{critical_states.dist_structure}
\alias{critical_states.dist_structure}
\alias{critical_states}
\title{Critical states of a component}
\usage{
\method{critical_states}{dist_structure}(x, j)

critical_states(x, j)
}
\arguments{
\item{x}{A \link{dist_structure} object.}

\item{j}{Component index.}
}
\value{
Integer matrix with \code{m - 1} columns; rows are the states of the
other components for which \code{j} is critical.
}
\description{
Critical states of a component
}
