% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/importance.R
\name{criticality_importance}
\alias{criticality_importance}
\alias{criticality_importance.dist_structure}
\title{Criticality (Fussell) importance at time t}
\usage{
criticality_importance(x, j, t)

\method{criticality_importance}{dist_structure}(x, j, t)
}
\arguments{
\item{x}{A \link{dist_structure} object.}

\item{j}{Component index.}

\item{t}{Scalar time.}
}
\value{
Numeric scalar in \verb{[0, 1]}.
}
\description{
Probability that component \code{j} has failed AND is critical given
that the system has failed by time \code{t}. Equals
\verb{I_B(j; S(t)) * F_j(t) / F_sys(t)}.
}
