% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_structure-class.R
\name{format.dist_structure}
\alias{format.dist_structure}
\title{Format a dist_structure object}
\usage{
\method{format}{dist_structure}(x, ...)
}
\arguments{
\item{x}{A \link{dist_structure} object.}

\item{...}{Ignored.}
}
\value{
Character vector suitable for \code{\link[=cat]{cat()}}.
}
\description{
Format a dist_structure object
}
