% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R, R/generics.R
\name{is_coherent.dist_structure}
\alias{is_coherent.dist_structure}
\alias{is_coherent}
\title{Coherence axiom check}
\usage{
\method{is_coherent}{dist_structure}(x)

is_coherent(x)
}
\arguments{
\item{x}{A \link{dist_structure} object.}
}
\value{
\code{TRUE} if monotone and every component relevant.
}
\description{
Coherence axiom check
}
