% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_structure-class.R
\name{is_dist_structure}
\alias{is_dist_structure}
\title{Predicate for dist_structure objects}
\usage{
is_dist_structure(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"dist_structure"}.
}
\description{
Predicate for dist_structure objects
}
