% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iid_constructors.R
\name{max_iid}
\alias{max_iid}
\title{Parallel system of m iid components}
\usage{
max_iid(d, m)
}
\arguments{
\item{d}{A \code{dist} object.}

\item{m}{Number of components.}
}
\value{
A \code{parallel_dist}.
}
\description{
Equivalent to the maximum of \code{m} iid random variables from \code{d}.
Preserves the parallel topology for structural queries.
}
