% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R, R/generics.R
\name{min_cuts.dist_structure}
\alias{min_cuts.dist_structure}
\alias{min_cuts}
\title{Minimal cut sets}
\usage{
\method{min_cuts}{dist_structure}(x)

min_cuts(x)
}
\arguments{
\item{x}{A \link{dist_structure} object.}
}
\value{
A list of integer vectors.
}
\description{
Minimal cut sets
}
