% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ncomponents}
\alias{ncomponents}
\title{Number of components}
\usage{
ncomponents(x)
}
\arguments{
\item{x}{A \link{dist_structure} object.}
}
\value{
A positive integer \code{m}, the number of components.
}
\description{
Number of components
}
