% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherent_dist.R
\name{parallel_dist}
\alias{parallel_dist}
\title{Parallel system distribution}
\usage{
parallel_dist(components)
}
\arguments{
\item{components}{List of \code{dist} objects.}
}
\value{
A \code{parallel_dist} inheriting from \code{coherent_dist}.
}
\description{
A parallel system fails only when all components fail. Equivalent to
\code{max(components)} but preserves topology.
}
