% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_structure-class.R
\name{print.dist_structure}
\alias{print.dist_structure}
\title{Print a dist_structure object}
\usage{
\method{print}{dist_structure}(x, ...)
}
\arguments{
\item{x}{A \link{dist_structure} object.}

\item{...}{Passed to \code{\link[=format]{format()}}.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a dist_structure object
}
