% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R, R/generics.R
\name{structural_importance.dist_structure}
\alias{structural_importance.dist_structure}
\alias{structural_importance}
\title{Birnbaum structural importance}
\usage{
\method{structural_importance}{dist_structure}(x, j)

structural_importance(x, j)
}
\arguments{
\item{x}{A \link{dist_structure} object.}

\item{j}{Component index.}
}
\value{
Numeric scalar in \verb{[0, 1]}.
}
\description{
Fraction of the \code{2^(m - 1)} states of the other components for which
component \code{j} is critical.
}
