% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R, R/generics.R
\name{system_censoring.dist_structure}
\alias{system_censoring.dist_structure}
\alias{system_censoring}
\title{Per-component censoring from system observation}
\usage{
\method{system_censoring}{dist_structure}(x, times)

system_censoring(x, times)
}
\arguments{
\item{x}{A \link{dist_structure} object.}

\item{times}{Non-negative numeric vector of length \code{ncomponents(x)}.}
}
\value{
A list with \code{system_time} (scalar) and \code{component_status}
(character vector of length \code{m}, values in \code{"exact"}, \code{"left"},
\code{"right"}).
}
\description{
Per-component censoring from system observation
}
