% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R, R/generics.R
\name{system_lifetime.dist_structure}
\alias{system_lifetime.dist_structure}
\alias{system_lifetime}
\title{System lifetime from component times}
\usage{
\method{system_lifetime}{dist_structure}(x, times)

system_lifetime(x, times)
}
\arguments{
\item{x}{A \link{dist_structure} object.}

\item{times}{Non-negative numeric vector of length \code{ncomponents(x)}.}
}
\value{
Scalar system lifetime.
}
\description{
System lifetime from component times
}
