% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/importance.R
\name{vesely_fussell_importance}
\alias{vesely_fussell_importance}
\alias{vesely_fussell_importance.dist_structure}
\title{Vesely-Fussell importance at time t}
\usage{
vesely_fussell_importance(x, j, t)

\method{vesely_fussell_importance}{dist_structure}(x, j, t)
}
\arguments{
\item{x}{A \link{dist_structure} object.}

\item{j}{Component index.}

\item{t}{Scalar time.}
}
\value{
Numeric scalar in \verb{[0, 1]}.
}
\description{
Probability that at least one minimal cut set containing \code{j} has all
its components failed, given the system has failed by time \code{t}.
Computed exactly via inclusion-exclusion over subsets of cuts that
contain \code{j}.
}
