% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherent_dist.R
\name{bridge_dist}
\alias{bridge_dist}
\title{Bridge system distribution}
\usage{
bridge_dist(components)
}
\arguments{
\item{components}{List of 5 \code{dist} objects.}
}
\value{
A \code{bridge_dist} inheriting from \code{coherent_dist}.
}
\description{
The classical 5-component bridge reliability network with minimal
path sets \verb{\{1,4\}}, \verb{\{2,5\}}, \verb{\{1,3,5\}}, \verb{\{2,3,4\}}.
}
