% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherent_dist.R
\name{kofn_dist}
\alias{kofn_dist}
\title{k-out-of-n system distribution}
\usage{
kofn_dist(k, components)
}
\arguments{
\item{k}{Minimum functioning components for system operation.}

\item{components}{List of \code{dist} objects (length \code{m}).}
}
\value{
A \code{kofn_dist} inheriting from \code{coherent_dist}.
}
\description{
A k-out-of-n system functions if at least \code{k} of its \code{m} components
function. Equivalent to the \code{(m - k + 1)}-th order statistic of
component lifetimes.
}
