% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R, R/generics.R
\name{min_paths.dist_structure}
\alias{min_paths.dist_structure}
\alias{min_paths}
\title{Minimal path sets}
\usage{
\method{min_paths}{dist_structure}(x)

min_paths(x)
}
\arguments{
\item{x}{A \link{dist_structure} object.}
}
\value{
A list of integer vectors.
}
\description{
Minimal path sets
}
