% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{min_paths}
\alias{min_paths}
\title{Minimal path sets}
\usage{
min_paths(x)
}
\arguments{
\item{x}{A \link{dist_structure} object.}
}
\value{
A list of integer vectors.
}
\description{
Minimal path sets
}
