% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R, R/generics.R
\name{phi.dist_structure}
\alias{phi.dist_structure}
\alias{phi}
\title{Structure function}
\usage{
\method{phi}{dist_structure}(x, state)

phi(x, state)
}
\arguments{
\item{x}{A \link{dist_structure} object.}

\item{state}{Integer or logical vector of length \code{ncomponents(x)} in
\verb{\{0, 1\}}.}
}
\value{
Integer scalar, \code{0} or \code{1}.
}
\description{
Evaluate the coherent structure function \verb{phi: \{0, 1\}^m -> \{0, 1\}} at a
component state vector \code{state}. By convention \code{state[j] = 1} means
component \code{j} is functioning; \code{phi(state) = 1} means the system is
functioning.
}
