% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{system_signature}
\alias{system_signature}
\title{System signature}
\usage{
system_signature(x)
}
\arguments{
\item{x}{A \link{dist_structure} object.}
}
\value{
Numeric vector of length \code{ncomponents(x)} summing to 1.
}
\description{
Samaniego's signature: \verb{s = (s_1, ..., s_m)} where \code{s_k} is the
probability the system fails at the k-th component failure under i.i.d.
component lifetimes.
}
